fix : delete button

+++ model/departmentlist.php
@@ -93,41 +93,47 @@ if (isset($_POST['edit_department']))
 if (isset($_POST['delete_department']))
 {
   $department_id = $_POST['department_id'];
-  $bulk = new MongoDB\Driver\BulkWrite;
-  $bulk->delete([
-                  '_id'=>new \MongoDB\BSON\ObjectID($department_id)], 
-                  ['limit' => 1]
-                );
-  $writeConcern = new MongoDB\Driver\WriteConcern(MongoDB\Driver\WriteConcern::MAJORITY, 1000);
-  try
+  $password = $_POST['password'];
+  $password_hash = $_SESSION["loggeduser_ConsumerPassword"];
+
+  if (password_verify($password, $password_hash))
   {
-    $result =$GoNGetzDatabase->executeBulkWrite('GoNGetzSmartSchool.SchoolsDepartment', $bulk, $writeConcern);
-  }
-  catch (MongoDB\Driver\Exception\BulkWriteException $e)
-  {
-    $result = $e->getWriteResult();
-    // Check if the write concern could not be fulfilled
-    if ($writeConcernError = $result->getWriteConcernError())
+    $bulk = new MongoDB\Driver\BulkWrite;
+    $bulk->delete([
+                    '_id'=>new \MongoDB\BSON\ObjectID($department_id)], 
+                    ['limit' => 1]
+                  );
+    $writeConcern = new MongoDB\Driver\WriteConcern(MongoDB\Driver\WriteConcern::MAJORITY, 1000);
+    try
     {
-        printf("%s (%d): %s\n",
-            $writeConcernError->getMessage(),
-            $writeConcernError->getCode(),
-            var_export($writeConcernError->getInfo(), true)
-        );
+      $result =$GoNGetzDatabase->executeBulkWrite('GoNGetzSmartSchool.SchoolsDepartment', $bulk, $writeConcern);
     }
-    // Check if any write operations did not complete at all
-    foreach ($result->getWriteErrors() as $writeError)
+    catch (MongoDB\Driver\Exception\BulkWriteException $e)
     {
-        printf("Operation#%d: %s (%d)\n",
-            $writeError->getIndex(),
-            $writeError->getMessage(),
-            $writeError->getCode()
-        );
+      $result = $e->getWriteResult();
+      // Check if the write concern could not be fulfilled
+      if ($writeConcernError = $result->getWriteConcernError())
+      {
+          printf("%s (%d): %s\n",
+              $writeConcernError->getMessage(),
+              $writeConcernError->getCode(),
+              var_export($writeConcernError->getInfo(), true)
+          );
+      }
+      // Check if any write operations did not complete at all
+      foreach ($result->getWriteErrors() as $writeError)
+      {
+          printf("Operation#%d: %s (%d)\n",
+              $writeError->getIndex(),
+              $writeError->getMessage(),
+              $writeError->getCode()
+          );
+      }
     }
-  }
-  catch (MongoDB\Driver\Exception\Exception $e)
-  {
-    printf("Other error: %s\n", $e->getMessage());
-    exit;
+    catch (MongoDB\Driver\Exception\Exception $e)
+    {
+      printf("Other error: %s\n", $e->getMessage());
+      exit;
+    }
   }
 }
